refactor: apply changes to tests
