+++ app/src/test/java/com/example/plantsapp/data/repository/RoomTasksRepositoryTest.kt
@@ -3,7 +3,7 @@ package com.example.plantsapp.data.repository
 import org.junit.Assert.assertEquals
 import com.example.plantsapp.data.dao.StubTasksDao
 import com.example.plantsapp.data.entity.RoomTask
-import com.example.plantsapp.data.entity.TaskKeys
+import com.example.plantsapp.domain.model.TaskKeys
 import com.example.plantsapp.data.utils.createPlant
 import com.example.plantsapp.data.utils.createRoomTask
 import com.example.plantsapp.domain.model.Task
